% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{as.double.CovEsts}
\alias{as.double.CovEsts}
\title{as.double Method for CovEsts Objects}
\usage{
\method{as.double}{CovEsts}(x, ...)
}
\arguments{
\item{x}{A CovEsts S3 object.}

\item{...}{Refer to \link[base:double]{base::double}.}
}
\value{
The acf values from the CovEsts object.
}
\description{
This function provides a method for as.double for CovEsts objects, allowing for as.numeric to be called.
It extracts the acf values from the object.
}
\examples{
as.numeric(standard_est(c(1, 2, 3)))
}
