% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{blackman_window}
\alias{blackman_window}
\title{Blackman Window.}
\usage{
blackman_window(x, alpha)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}

\item{alpha}{A real number which controls the height of the spike. A negative alpha creates a positive peak and a positive alpha creates a negative peak.}
}
\value{
A vector of values.
}
\description{
This computes the Blackman window.
\deqn{w(x) = ( (1 - \alpha) / 2) - (1/2) \cos(\pi x) + (\alpha / 2) \cos(2 \pi x), x \in [0, 1], \alpha \in \mathbb{R}.}
As \eqn{\alpha} increases, the function
}
\examples{
X <- c(0.1, 0.5, 0.75)
blackman_window(X, 1.34)
}
