% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Title}
\usage{
bootstrap(X, l, k)
}
\arguments{
\item{X}{A vector representing observed values of the time series.}

\item{l}{The block length considered for bootstrap.}

\item{k}{The number of blocks considered for bootstrap.}
}
\description{
Title
}
\examples{
X <- c(1, 2, 3, 3, 2, 1)
bootstrap(X, 2, 3)
}
