% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_1d_dct}
\alias{compute_1d_dct}
\title{Compute 1D Discrete Cosine Transform}
\usage{
compute_1d_dct(X)
}
\arguments{
\item{X}{A vector of values for which the discrete cosine transform is being computed.}
}
\value{
A vector whose values are the transform.
}
\description{
This computes the Type-II discrete cosine transform.
}
\examples{
X <- c(1, 2, 3)
compute_1d_dct(X)
}
