% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_1d_idct}
\alias{compute_1d_idct}
\title{Compute 1D Inverse Discrete Cosine Transform}
\usage{
compute_1d_idct(X)
}
\arguments{
\item{X}{A vector of values for which the discrete cosine transform is being computed.}
}
\value{
A vector whose values are the inverse transform.
}
\description{
This computes the inverse of the Type-II discrete cosine transform.
}
\examples{
X <- compute_1d_dct(c(1, 2, 3))
compute_1d_idct(X)
}
