% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_adjusted}
\alias{compute_adjusted}
\title{Compute the Kernel Regression Estimator.}
\usage{
compute_adjusted(
  X,
  x,
  meanX,
  t,
  h,
  kernel = "gaussian",
  kernel_params = c(),
  custom_kernel = FALSE,
  pd = TRUE
)
}
\arguments{
\item{X}{A vector of values representing a stochastic process.}

\item{x}{A vector of indices.}

\item{meanX}{The mean of the process \eqn{X.}}

\item{t}{The values at which the covariance function is calculated at.}

\item{h}{Bandwidth parameter.}

\item{kernel}{The name of the kernel function to be used. Possible values are:
"gaussian", "wave", "rational_quadratic", and "bessel_j".}

\item{kernel_params}{Any parameters for the kernel function.}

\item{custom_kernel}{A boolean determining whether or not a custom kernel is used.}

\item{pd}{Whether or not the estimator returned is positive-definite.}
}
\value{
A vector whose values are the estimated kernel regression estimator.
}
\description{
\deqn{
\hat{\rho}(t) = \left( \sum_{i} \sum_{j} \widecheck{X}_{ij} K((t - t_{ij}) / h) \right) \left( \sum_{i} \sum_{j} K((t - t_{ij}) / h) \right)^{-1},
}
where \eqn{\widecheck{X}_{ij} = (X(t_{i}) - \overline{X}) (X(t_{j}) - \overline{X})}, and \eqn{t_{ij} = t_{i} - t_{j}.}
}
\details{
When making the estimator positive-definite, the following is done:
\enumerate{
\item Take the discrete cosine transform.
\item Set all negative values to zero.
\item Perform the inversion.
This ensures the covariance estimate is positive-definite.
}
}
\examples{
X <- c(1, 2, 3, 4)
compute_adjusted(X, 1:4, mean(X), 1:3, 0.1, "gaussian", c(), FALSE, TRUE)
}
