% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapered_estimator.R
\name{compute_h2n}
\alias{compute_h2n}
\title{Compute \eqn{H_{2, n}}.}
\usage{
compute_h2n(n, rho, window_name, window_params = c(1), custom_window = FALSE)
}
\arguments{
\item{n}{The number of samples.}

\item{rho}{A scale paramter, greater than 0.}

\item{window_name}{The window function to use, one of
"tukey", "triangular", "power_sine", "blackman_window", "hann_poisson", "welch".}

\item{window_params}{Any parameters of the window function.}

\item{custom_window}{Whether or not a custom window is used (boolean). Put its parameters in window_params.}
}
\value{
A single value being \eqn{H_{2, n}(0)}.
}
\description{
This is used in the computation of \link{compute_tapered_cov_single}.
\deqn{H_{2, n}(0) = \sum_{s=1}^{n} a((s - 1/2) / n)^2 , }
where \eqn{a} is a window function.
}
\examples{
compute_h2n(3, 0.6, "tukey")
}
