% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_standard_estimator.R
\name{compute_kernel_corrected_est}
\alias{compute_kernel_corrected_est}
\title{Kernel correction for an estimated autocovariance function.}
\usage{
compute_kernel_corrected_est(
  cov,
  maxLag,
  kernel_name,
  kernel_params = c(),
  N_T = 0.1 * length(cov),
  custom_kernel = FALSE
)
}
\arguments{
\item{cov}{A vector whose values are an estimate autocovariance function.}

\item{maxLag}{The maximum lag to compute the autocovariance function at.}

\item{kernel_name}{The name of the kernel function to be used. Possible values are:
"gaussian", "exponential", "wave", "rational_quadratic", "spherical", "circular", "bessel_j", "matern", "cauchy".}

\item{kernel_params}{A vector of parameters of the kernel function. See \link{kernel} for parameters.
In the case of "gaussian", "wave", "rational_quadratic", "spherical" and "circular", no parameter is passed as \eqn{\theta} is \code{N_T}.}

\item{N_T}{The rate at which the kernel function vanishes at. Recommended to be \eqn{0.1 N} when considering all lags. This parameter may be large for a small range of estimation lags.}

\item{custom_kernel}{If a custom kernel is to be used or not.  See the examples of \link{compute_corrected_standard_est} for usage.}
}
\value{
A vector whose values are a kernel corrected autocovariance estimator.
}
\description{
Kernel correction for an estimated autocovariance function.
}
\examples{
X <- rnorm(1000)
Y <- c(X[1], X[2])
for(i in 3:length(X)) { Y[i] <- X[i] - 0.3*X[i - 1] - 0.6*X[i - 2] }
cov_est <- compute_standard_est(Y, length(Y) - 1)
plot(compute_kernel_corrected_est(cov_est, length(Y)-1,
     "bessel_j", kernel_params=c(0, 1), N_T=0.2*length(Y)))
}
