% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_rho_T1}
\alias{compute_rho_T1}
\title{Compute \eqn{\rho(T1)} for the truncated kernel regression estimator.}
\usage{
compute_rho_T1(
  x,
  meanX,
  T1,
  h,
  Xij_mat,
  kernel = "gaussian",
  kernel_params = c(),
  custom_kernel = FALSE
)
}
\arguments{
\item{x}{A vector of indices.}

\item{meanX}{The mean of the process \eqn{X.}}

\item{T1}{The first trunctation point.}

\item{h}{Bandwidth parameter.}

\item{Xij_mat}{The matrix of pairwise covariance values.}

\item{kernel}{The name of the kernel function to be used. Possible values are:
"gaussian", "wave", "rational_quadratic", and "bessel_j".}

\item{kernel_params}{Any parameters for the kernel function.}

\item{custom_kernel}{A boolean determining whether or not a custom kernel is used.}
}
\value{
The empirical covariance function a T1.
}
\description{
This function computes the truncated kernel regression estimator at \eqn{\rho(T_{1}).}
This saves repeated computations in compute_truncated.
}
\examples{
X <- c(1, 2, 3, 4)
compute_rho_T1(1:4, mean(X), 1, 0.1, compute_Xij_mat(X), "gaussian", c(), FALSE)
}
\references{
Hall, P., & Patil, P. (1994). Properties of nonparametric estimators of autocovariance for stationary random fields. In Probability Theory and Related Fields (Vol. 99, Issue 3, pp. 399–424). 10.1007/bf01199899
Hall, P., Fisher, N. I., & Hoffmann, B. (1994). On the Nonparametric Estimation of Covariance Functions. In The Annals of Statistics (Vol. 22, Issue 4). Institute of Mathematical Statistics. 10.1214/aos/1176325774
}
