% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_standard_estimator.R
\name{compute_sandwich_est}
\alias{compute_sandwich_est}
\title{Compute the Long-Run Covariance Sandwich Estimator}
\usage{
compute_sandwich_est(
  X,
  window_name = "tukey",
  window_params = c(1),
  b = floor(length(X)^(1/5)),
  pd = TRUE,
  maxLag = length(X) - 1,
  meanX = mean(X),
  custom_window = FALSE
)
}
\arguments{
\item{X}{A vector representing observed values of the time series.}

\item{window_name}{The name of the \link{window} function to be used. Possible values are:
tukey, triangular, power_sine, blackman_window, hann_poisson, welch. Alternatively, a custom window function can be provided, see the example}

\item{window_params}{A vector of parameters of the window function.}

\item{b}{Bandwidth parameter, greater than 0. Defaults to \eqn{\lfloor} length(X) \eqn{\rfloor^{1/5}.}}

\item{pd}{Whether the standard positive-definite estimate should be used. Defaults to \code{TRUE}.}

\item{maxLag}{An optional parameter that determines the maximum lag to compute the estimated long-run covariance. Defaults to \code{length(X) - 1}.}

\item{meanX}{The average value of \code{X}. Defaults to \code{mean(X)}.}

\item{custom_window}{If a custom window is to be used or not. Defaults to \code{FALSE}.}
}
\value{
The long-run covariance of the time series.
}
\description{
This function computes the long-run covariance sandwich estimator of a time series, proposed in Shang (2024).
}
\details{
This function computes the long-run covariance sandwich estimator of a time series.
The long-run covariance is a measure of total dependence in a time series as opposed to at a single lag, like the autocovariance.

Shang (2024) provides the following estimator (adapted for the nonfunctional time series case)
\deqn{
\widehat{C}_{b} = \sum_{h=-(N-1)}^{N-1} w(h / b) \widehat{C}(h) ,
}
where \eqn{b > 0} is the bandwith parameter, \eqn{w(\cdot)} is a symmetric window function such that \eqn{w(0) = 1, w(\cdot) \leq 1} and after some value \eqn{\iota} (we selected 1) \eqn{w(\cdot) = 0}.
In the paper, \eqn{\widehat{C}(\cdot)} is the standard positive-definite autocovariance estimator (see \link{standard_est}).
}
\examples{
X <- 0:4
compute_sandwich_est(X)
my_window <- function(x, ...) {
  retVal <- 1 - x
  retVal[abs(x) > 1] <- 0
  return(retVal)
}
compute_sandwich_est(X, window_name=my_window, custom_window = TRUE)
}
\references{
Shang, H. L. (2024). Bootstrapping Long-Run Covariance of Stationary Functional Time Series. Forecasting 6(1), 138-151. 10.3390/forecast6010008
}
