% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapered_estimator.R
\name{compute_taper}
\alias{compute_taper}
\title{Compute the function \eqn{a(x; \rho).}}
\usage{
compute_taper(x, rho, window_name, window_params = c(1), custom_window = FALSE)
}
\arguments{
\item{x}{A vector of number between 0 and 1 (inclusive).}

\item{rho}{A scale paramter, greater than 0.}

\item{window_name}{The window function to use, one of
"tukey", "triangular", "power_sine", "blackman_window", "hann_poisson", "welch".}

\item{window_params}{Any parameters of the window function.}

\item{custom_window}{Whether or not a custom window is used (boolean). Put its parameters in window_params.}
}
\value{
A vector of taper values.
}
\description{
This function repeatedly calls \link{compute_taper_single}.
}
\examples{
X <- c(0.1, 0.2, 0.3)
compute_taper(X, 0.5, "tukey")
}
