% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapered_estimator.R
\name{compute_tapered_cov}
\alias{compute_tapered_cov}
\title{Compute the tapered covariance function over a set of lags.}
\usage{
compute_tapered_cov(
  X,
  maxLag,
  rho,
  window_name,
  window_params = c(1),
  custom_window = FALSE
)
}
\arguments{
\item{X}{A numeric vector representing the process.}

\item{maxLag}{The maximum lag at which the covariance function is to be computed. Must be less than the length of the process.}

\item{rho}{A scale parameter greater than zero.}

\item{window_name}{The window function to use, one of
"tukey", "triangular", "power_sine", "blackman_window", "hann_poisson", "welch". Alternatively, a custom window function can be provided, see \link{compute_taper_single}'s example.}

\item{window_params}{Any parameters of the window function.}

\item{custom_window}{Whether or not a custom window is used (boolean). Put its parameters in window_param}
}
\value{
A numeric vector representing values of the tapered covariance function.
}
\description{
This computes the tapered covariance over a set of lags,
\deqn{\widehat{C}_{N}^{a} (h) = (H_{2, n}(0))^{-1} \sum_{t=1}^{n-h}  (X(t_{i}) - \bar{X} ) ( X(t_{i} + h) - \bar{X} ) [ a((t_{i} - 1/2) / n; \rho) a((t_{i} + h - 1/2) / n; \rho)  ] ,}
where \eqn{a} is a window function, \eqn{\rho} is a scale parameter.
For each lag, the tapered covariance is computed in \link{compute_tapered_cov_single}.
}
\details{
This computes the tapered covariance over a set of lags,
\deqn{\widehat{C}_{N}^{a} (h) = (H_{2, n}(0))^{-1} \sum_{t=1}^{n-h}  (X(t_{i}) - \bar{X} ) ( X(t_{i} + h) - \bar{X} ) [ a((t_{i} - 1/2) / n; \rho) a((t_{i} + h - 1/2) / n; \rho)  ] ,}
where \eqn{a} is a window function, \eqn{\rho} is a scale parameter.
This estimator considers the edge effect during estimation, assigning a lower weight to values closer to the edges and higher weights for observations closer to the middle.
This estimator is positive-definite and asymptotically unbiased.

Internally, this function calls \link{compute_tapered_cov_single} for each lag \eqn{h}.

The first set of taper values (\eqn{a((t_{i} - 1/2) / n; \rho)}) are precomputed in order to save time, as they do not vary,
and the second set (\eqn{a((t_{i} + h - 1/2) / n; \rho)}) are computed once for each lag \eqn{h}.
}
\examples{
X <- c(1, 2, 3)
compute_tapered_cov(X, 2, 0.5, "tukey")
}
\references{
Dahlhaus R., Künsch, H. (1987). Edge Effects and Efficient Parameter Estimation feor Stationary Random Fields. In Biometrika (Vol. 74 Issue 4, pp. 877-882). 10.1093/biomet/74.4.877
}
