% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_truncated}
\alias{compute_truncated}
\title{Compute the Truncated Kernel Regression Estimator.}
\usage{
compute_truncated(
  X,
  x,
  meanX,
  t,
  T1,
  T2,
  h,
  kernel_name = "gaussian",
  kernel_params = c(),
  custom_kernel = FALSE,
  pd = TRUE
)
}
\arguments{
\item{X}{A vector of values representing a stochastic process.}

\item{x}{A vector of indices.}

\item{meanX}{The mean of the process \eqn{X.}}

\item{t}{The values at which the covariance function is calculated at.}

\item{T1}{The first truncation point, \eqn{T_{1} > 0.}}

\item{T2}{The second truncation point, \eqn{T_{2} > T_{1} > 0.}}

\item{h}{Bandwidth parameter.}

\item{kernel_name}{The name of the kernel function to be used. Possible values are:
"gaussian", "wave", "rational_quadratic", and "bessel_j". Alternatively, a custom kernel function can be provided, see \link{compute_corrected_standard_estimator}'s example.}

\item{kernel_params}{Any parameters for the kernel function.}

\item{custom_kernel}{A boolean determining whether or not a custom kernel is used.}

\item{pd}{Whether or not the estimator returned is positive-definite.}
}
\value{
A vector whose values are the truncated kernel regression estimator.
}
\description{
This function computes the truncated kernel regression estimator,
\deqn{\hat{\rho}_{1}(t)  = \left\{ \begin{array}{ll}
\hat{\rho}(t) & 0 \leq t \leq T_{1} \\
\hat{\rho}(T_{1}) (T_{2} - t)(T_{2} - T_{1})^{-1} & T_{1} < t \leq T_{2} \\
0 & t > T_{2}
\end{array} , \right. }
where \eqn{\hat{\rho}(\cdot)} is the kernel regression estimator, see \link{compute_adjusted}.
}
\details{
This function computes the truncated kernel regression estimator,
\deqn{\hat{\rho}_{1}(t) = \left\{ \begin{array}{ll}
\hat{\rho}(t) & 0 \leq t \leq T_{1} \\
\hat{\rho}(T_{1}) (T_{2} - t)(T_{2} - T_{1})^{-1} & T_{1} < t \leq T_{2} \\
0 & t > T_{2}
\end{array} \right. }
where \eqn{\hat{\rho}(\cdot)} is the kernel regression estimator, see \link{compute_adjusted}.

Compared to \link{compute_adjusted}, this function brings down the estimate to zero linearly between \eqn{T_{1}} and \eqn{T_{2}}.
In the case of short-range dependence, this may be beneficial as it can remove estimation artefacts.
a
To make this estimator positive-definite, the following procedure is done:
\enumerate{
\item Take the discrete cosine transform
\eqn{\mathcal{F}^{c}(\hat{\rho}_{1}(t))}.
\item Find the smallest frequency where its associated value in the spectral domain is negative
\deqn{\hat{\theta} = \inf\{ \theta > 0 :  \mathcal{F}^{c}(\hat{\rho}_{1}(t)) < 0\}.}
\item Set all values starting at the frequency to zero.
\item Perform the inversion.
}

This ensures the covariance function estimate is positive-definite.
}
\examples{
X <- c(1, 2, 3, 4)
compute_truncated(X, 1:4, mean(X), 1:3, 1, 2, 0.1,
                  "gaussian", c(), FALSE, TRUE)
}
\references{
Hall, P., & Patil, P. (1994). Properties of nonparametric estimators of autocovariance for stationary random fields. In Probability Theory and Related Fields (Vol. 99, Issue 3, pp. 399–424). 10.1007/bf01199899

Hall, P., Fisher, N. I., & Hoffmann, B. (1994). On the Nonparametric Estimation of Covariance Functions. In The Annals of Statistics (Vol. 22, Issue 4). Institute of Mathematical Statistics. 10.1214/aos/1176325774
}
