% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_Xij_mat}
\alias{compute_Xij_mat}
\title{Compute \eqn{X_{ij}} Matrix}
\usage{
compute_Xij_mat(X)
}
\arguments{
\item{X}{A vector of values}
}
\value{
A matrix of size \eqn{N \times N}, where \eqn{N} is the length of the vector X.
}
\description{
This computes the matrix of pairwise covariance values, \eqn{X_{ij},} for the kernel regression estimator,
\deqn{X_{ij} = (X_{i} - \bar{X}) (X_{j} - \bar{X}) .}
}
\examples{
X <- c(1, 2, 3, 4)
compute_Xij_mat(X)
}
\references{
Hall, P., & Patil, P. (1994). Properties of nonparametric estimators of autocovariance for stationary random fields. In Probability Theory and Related Fields (Vol. 99, Issue 3, pp. 399–424). 10.1007/bf01199899
Hall, P., Fisher, N. I., & Hoffmann, B. (1994). On the Nonparametric Estimation of Covariance Functions. In The Annals of Statistics (Vol. 22, Issue 4). Institute of Mathematical Statistics. 10.1214/aos/1176325774
}
