% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_standard_estimator.R
\name{corrected_standard_estimator}
\alias{corrected_standard_estimator}
\title{Kernel correction of the standard estimator.}
\usage{
corrected_standard_estimator(
  X,
  upperTau,
  kernel,
  kernel_params = c(),
  N_T = 0.1 * length(X),
  N = length(X),
  meanX = mean(X),
  pd = TRUE,
  type = "covariance",
  customKernel = FALSE
)
}
\arguments{
\item{X}{A vector of observations representing a time series.}

\item{upperTau}{The maximum upper lag to compute the autocovariance function. upperTau \eqn{< N-1.}}

\item{kernel}{The choice of kernel. Possible values are:
"gaussian", "exponential", "wave", "rational_quadratic", "spherical", "circular", "bessel_j", "matern", "cauchy".}

\item{kernel_params}{If the kernel has any parameters, pass them as a vector.}

\item{N_T}{The value the kernel function vanishes at. Recommended to be \eqn{0.1 N.}}

\item{N}{The length of the vector X.}

\item{meanX}{The average value of the observations.}

\item{pd}{Whether a positive definite estimate should be used.}

\item{type}{Whether the autocovariance or autocorrelation should be computed, options: 'covariance', 'correlation'.}

\item{customKernel}{Whether or not you want to use a custom kernel function. This will allow you to define your own kernel function. See the examples for usage.}
}
\value{
The autocovariance up to lag upperTau for the time series X.
}
\description{
This applies kernel correction to the standard estimators.
It considers a kernel \eqn{a(\cdot)} which vanishes after a certain distancem, \eqn{N_{T}}, which is recommended to be \eqn{0.1 N}, where \eqn{N} is the length of the process.
\deqn{\widehat{C}^{(a)}(h) = \widehat{C}(h) a(h / N_{T}).}
Also, \eqn{a(0) = 1.}
}
\examples{
X <- c(1, 2, 3)
corrected_standard_estimator(X, 2, "gaussian")

X <- rnorm(1000)
Y <- c(X[1], X[2])
for(i in 3:length(X)) { Y[i] <- X[i] - 0.3*X[i - 1] - 0.6*X[i - 2] }
plot(Y)
plot(corrected_standard_estimator(Y, length(Y)-1,
     "bessel_j", kernel_params=c(0, 1), N_T=0.2*length(Y)))

# Custom kernel
my_kernel <- function(x, theta, params) {
  stopifnot(theta > 0, length(x) >= 1, all(x >= 0))
  return(sapply(x, function(t) ifelse(t == 0, 1, ifelse(t == Inf, 0, (sin((t^params[1]) / theta) / ((t^params[1]) / theta)) * cos((t^params[2]) / theta)))))
}
X <- rnorm(1000)
Y <- c(X[1], X[2])
for(i in 3:length(X)) { Y[i] <- X[i] - 0.3*X[i - 1] - 0.6*X[i - 2] }
plot(Y)
plot(corrected_standard_estimator(Y, length(Y)-1, "my_kernel", kernel_params=c(2, 0.25), customKernel = TRUE))
}
\references{
Yaglom AM (1987). Correlation Theory of Stationary and Related Random Functions. Volume I: Basic Results. Springer New York. 10.1007/978-1-4612-4628-2.
}
