% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{CovEsts-deprecated}
\alias{CovEsts-deprecated}
\alias{kernel}
\alias{kernel_symm}
\alias{window}
\alias{window_symm}
\title{Deprecated functions in package \pkg{CovEsts}.}
\usage{
kernel(x, name, params = c(1))

kernel_symm(x, name, params = c(1))

window(x, name, params = c(1))

window_symm(x, name, params = c(1))
}
\arguments{
\item{x, name, params}{Parameters passed to the updated functions.}
}
\description{
These functions are provided for compatibility with older verisons of the \pkg{CovEsts} package.
These functions may be removed entirely.
}
\section{\code{kernel}}{

For \code{kernel()}, use \code{\link{kernel_ec}} instead.
}

\section{\code{kernel_symm}}{

For \code{kernel_symm()}, use \code{\link{kernel_symm_ec}} instead.
}

\section{\code{window}}{

For \code{window()}, use \code{\link{window_ec}} instead.
}

\section{\code{window_symm}}{

For \code{window_symm()}, use \code{\link{window_symm_ec}} instead.
}

\keyword{internal}
