% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_splines_est.R
\name{f_j_l}
\alias{f_j_l}
\title{Compute \eqn{f_{j}^{[l]}(x)}.}
\usage{
f_j_l(x, j, l, p, m, taus)
}
\arguments{
\item{x}{Numeric value for which the function is calculated at.}

\item{j}{Index of basis function of order \eqn{l}.}

\item{l}{Order of function.}

\item{p}{The order of the splines.}

\item{m}{The number of nonboundary knots.}

\item{taus}{Vector of \eqn{\tau}s, see \link{get_tau}.}
}
\value{
A numeric value of \eqn{f_{j}^{[l]}(x).}
}
\description{
This is an implementation of the formula in Choi, Li & Wang (2013, p. 616).
\deqn{
f_{j}^{[l]}(x) = \frac{m + 1}{l} \left( f_{j}^{[l - 1]}(x + 1) - \tau_{j - p} f_{j}^{[l - 1]}(x) + \tau_{j - p + l + 1} f_{j + 1}^{[l - 1]}(x) -  f_{j + 1}^{[l - 1]}(x + 1)  \right)
}
}
\examples{
taus <- get_all_tau(3, 2)
f_j_l(1, 2, 1, 3, 2, taus)
}
\references{
Choi, I., Li, B. & Wang, X. (2013). Nonparametric Estimation of Spatial and Space-Time Covariance Function. JABES (Vol. 18, pp. 611–630). 10.1007/s13253-013-0152-z
}
