% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapered_estimator.R
\name{H2n}
\alias{H2n}
\title{Compute \eqn{H_{2, n}}.}
\usage{
H2n(n, rho, window_name, window_params = c(1), custom_window = FALSE)
}
\arguments{
\item{n}{The number of samples.}

\item{rho}{A scale paramter, greater than 0.}

\item{window_name}{The name of the window function to be used. Possible values are:
"tukey", "triangular", "power_sine", "blackman_window", "hann_poisson", "welch". Alternatively, a custom window function can be provided, see the example.}

\item{window_params}{A vector of parameters of the window function.}

\item{custom_window}{If a custom window is to be used or not.}
}
\value{
A single value being \eqn{H_{2, n}(0)}.
}
\description{
This function is used in the computation of \link{tapered_cov_single}.
\deqn{H_{2, n}(0) = \sum_{s=1}^{n} a((s - 1/2) / n)^2 , }
where \eqn{a} is a window function.
}
\examples{
H2n(3, 0.6, "tukey")
}
