% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{hann_poisson_window}
\alias{hann_poisson_window}
\title{Hann-Poisson Window.}
\usage{
hann_poisson_window(x, alpha, ...)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}

\item{alpha}{A real number which controls the height of the spike. A negative alpha creates a large positive peak, try to avoid negative values above \eqn{-5} unless otherwise needed.}

\item{...}{Extra unusued arguments.}
}
\value{
A vector of values.
}
\description{
This computes the Hann-Poisson window.
\deqn{w(x) = (1/2) (1 - \cos(\pi x)) \exp( - (\alpha \left|1 - x \right|) ) , x \in [0, 1], \alpha \in \mathbb{R} .}
}
\examples{
X <- c(0.1, 0.5, 0.75)
hann_poisson_window(X, 0.2)
}
