% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{idct_1d}
\alias{idct_1d}
\title{Compute 1D Inverse Discrete Cosine Transform}
\usage{
idct_1d(X)
}
\arguments{
\item{X}{A vector of values for which the discrete cosine transform is being computed.}
}
\value{
A vector with the inverse transform values.
}
\description{
This helper function computes the inverse of the Type-II discrete cosine transform.
}
\details{
The Type-II inverse discrete cosine transform is computed using \link[stats:fft]{stats::fft}.
The original spectrum is obtained from the input X and then an inverse FFT is applied to obtain
the estimate.

The original spectrum, \code{dct_full}, from \code{X} is obtained as follows:
\code{dct_full <- c(X, 0, -X[-1], 0, rev(X[-1]))}. After this, an inverse FFT is applied,
\code{idct <- Re(stats::fft(dct_full, inverse = T)) * (2 / length(dct_full))}, and then the first half of inverse transform is taken, and the zeroes are dropped,
which gives the untransformed X.
}
\examples{
X <- dct_1d(c(1, 2, 3))
idct_1d(X)
}
\references{
Wikipedia contributors. (2025, January 24). Discrete cosine transform. In Wikipedia, The Free Encyclopedia. https://en.wikipedia.org/w/index.php?title=Discrete_cosine_transform&oldid=1271621657
}
