% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_cauchy}
\alias{kernel_cauchy}
\title{Isotropic Cauchy Kernel.}
\usage{
kernel_cauchy(x, theta, alpha, beta)
}
\arguments{
\item{x}{A vector of values of at least length 1.}

\item{theta}{Scale parameter, \eqn{\theta > 0.}}

\item{alpha}{A parameter for the kernel, \eqn{0 < a \leq 2.}}

\item{beta}{A parameter for the kernel, \eqn{b > 0.}}
}
\value{
A vector of values.
}
\description{
This computes the isotropic Cauchy kernel, which is valid for \eqn{\mathbb{R}^{d}} for \eqn{0 < \alpha \leq 2} and \eqn{\beta \geq 0.}
\deqn{a(\tau ; \theta, \alpha, \beta) = (1 + (\tau / \theta)^{\alpha})^{-(\beta / \alpha)} .}
}
\examples{
x <- 1
kernel_cauchy(x, 2, 1, 1)
x <- c(1, 2, 3)
kernel_cauchy(x, 1, 2, 3)
}
\references{
Table 4.2 of Hristopulos, D. T. (2020). Random Fields for Spatial Data Modeling:  A Primer for Scientists and Engineers. In Advances in Geographic Information Science. Springer Netherlands. https://doi.org/10.1007/978-94-024-1918-4
}
