% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_standard_estimator.R
\name{kernel_corrected_estimator}
\alias{kernel_corrected_estimator}
\title{Kernel correction for an estimated autocovariance function.}
\usage{
kernel_corrected_estimator(
  cov,
  upperTau,
  kernel_name,
  kernel_params = c(),
  N_T = 0.1 * length(cov),
  customKernel = FALSE
)
}
\arguments{
\item{cov}{A vector whose values are an estimate autocovariance function.}

\item{upperTau}{The maximum upper lag to compute the autocovariance function. upperTau \eqn{< N-1,} where \eqn{N} is the length of \code{cov}.}

\item{kernel_name}{The choice of kernel. Possible values are:
"gaussian", "exponential", "wave", "rational_quadratic", "spherical", "circular", "bessel_j", "matern", "cauchy".}

\item{kernel_params}{If the kernel has any parameters, pass them as a vector. See \link{kernel} for parameters.
In the case of "gaussian", "wave", "rational_quadratic", "sphericall" and "circular", no parameter is passed as \eqn{\theta} is \code{N_T}.}

\item{N_T}{The value the kernel function vanishes at. Recommended to be \eqn{0.1 N} when considering all lags. This parameter may be large for a small range of estimation lags.}

\item{customKernel}{Whether or not you want to use a custom kernel function. This will allow you to define your own kernel function. See the examples of \link{corrected_standard_estimator} for usage.}
}
\value{
A vector whose values are the estimated autocovariance up to lag upperTau.
}
\description{
Kernel correction for an estimated autocovariance function.
}
\examples{
X <- rnorm(1000)
Y <- c(X[1], X[2])
for(i in 3:length(X)) { Y[i] <- X[i] - 0.3*X[i - 1] - 0.6*X[i - 2] }
cov_est <- standard_est(Y, length(Y) - 1)
plot(kernel_corrected_estimator(Y, length(Y)-1,
     "bessel_j", kernel_params=c(0, 1), N_T=0.2*length(Y)))
}
