% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_gaussian}
\alias{kernel_gaussian}
\title{Isotropic Gaussian Kernel.}
\usage{
kernel_gaussian(x, theta)
}
\arguments{
\item{x}{A vector of values of at least length 1.}

\item{theta}{Scale parameter, \eqn{\theta > 0.}}
}
\value{
A vector of values.
}
\description{
This computes the isotropic Gaussian kernel, which is valid for \eqn{\mathbb{R}^{d}.}
\deqn{a(\tau;\theta) = \exp(-\tau^{2} / \theta).}
}
\examples{
x <- 1
kernel_gaussian(x, 1)
x <- c(1, 2, 3)
kernel_gaussian(x, 0.2)
}
\references{
Genton, Marc. (2001). Classes of Kernels for Machine Learning: A Statistics Perspective. Journal of Machine Learning Research. 2. 299-312. 10.1162/15324430260185646.
}
