% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_matern}
\alias{kernel_matern}
\title{Isotropic Matérn Kernel.}
\usage{
kernel_matern(x, theta, nu)
}
\arguments{
\item{x}{A vector of values of at least length 1.}

\item{theta}{Scale parameter, \eqn{\theta > 0.}}

\item{nu}{A parameter \eqn{\nu > 0.}}
}
\value{
A vector of values.
}
\description{
This computes the isotropic Matérn kernel, which is valid for \eqn{\mathbb{R}^{d}} for \eqn{\nu > 0.}
\deqn{a(\tau; \nu) = \left(\sqrt{2\nu} \frac{\tau}{\theta} \right)^{\nu} \left(2^{\nu - 1} \Gamma(\nu) \right)^{-1} K_{\nu}\left( \sqrt{2\nu} \frac{\tau}{\theta} \right) ,}
where \eqn{K_{\nu}(\cdot)} is the modified Bessel function of the second kind.
}
\examples{
x <- 1
kernel_matern(x, 2, 1)
x <- c(1, 2, 3)
kernel_matern(x, 0.5, 2)
}
\references{
Table 4.2 of Hristopulos, D. T. (2020). Random Fields for Spatial Data Modeling:  A Primer for Scientists and Engineers. In Advances in Geographic Information Science. Springer Netherlands. https://doi.org/10.1007/978-94-024-1918-4
Genton, Marc. (2001). Classes of Kernels for Machine Learning: A Statistics Perspective. Journal of Machine Learning Research. 2. 299-312. 10.1162/15324430260185646.
}
