% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_spherical}
\alias{kernel_spherical}
\title{Isotropic Spherical Kernel.}
\usage{
kernel_spherical(x, theta)
}
\arguments{
\item{x}{A vector of values of at least length 1.}

\item{theta}{Scale parameter, \eqn{\theta > 0.}}
}
\value{
A vector of values.
}
\description{
This computes the isotropic spherical kernel, which is valid for \eqn{\mathbb{R}^{3}.}
\deqn{a(\tau;\theta) =
\begin{cases}
1 - \frac{3}{2}\frac{\tau}{\theta} + \frac{1}{2}\left( \frac{\tau}{\theta} \right)^{3}, & \tau < \theta \\
0 & \text{otherwise}
\end{cases} .
}
}
\examples{
x <- 1
kernel_spherical(x, 2)
x <- c(1, 2, 3)
kernel_spherical(x, 2)
}
\references{
Genton, Marc. (2001). Classes of Kernels for Machine Learning: A Statistics Perspective. Journal of Machine Learning Research. 2. 299-312. 10.1162/15324430260185646.
}
