% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_kernels.R
\name{kernel_symm_gaussian}
\alias{kernel_symm_gaussian}
\title{Symmetric Gaussian Kernel.}
\usage{
kernel_symm_gaussian(x, theta, ...)
}
\arguments{
\item{x}{A vector of values of at least length 1.}

\item{theta}{Scale parameter, \eqn{\theta > 0.}}

\item{...}{Extra arguments. Unused.}
}
\value{
A vector of values.
}
\description{
This computes the Gaussian kernel. It has the properties of a probability distribution, unlike \code{kernel_gaussian(x, theta)}.
\deqn{a(\tau;\theta) = \sqrt{\pi \theta} \exp(-\tau^{2} / \theta).}
}
\examples{
x <- 1
kernel_symm_gaussian(x, 1)
x <- c(-3, -2, -1, 0, 1, 2, 3)
kernel_symm_gaussian(x, 0.2)
}
\references{
Genton, Marc. (2001). Classes of Kernels for Machine Learning: A Statistics Perspective. Journal of Machine Learning Research. 2. 299-312. 10.1162/15324430260185646.
}
