% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_wave}
\alias{kernel_wave}
\title{Isotropic Wave (Cardinal Sine) Kernel.}
\usage{
kernel_wave(x, theta)
}
\arguments{
\item{x}{A vector of values of at least length 1.}

\item{theta}{Scale parameter, \eqn{\theta > 0.}}
}
\value{
A vector of values.
}
\description{
This computes the isotropic wave (cardinal sine) kernel, which is valid for \eqn{\mathbb{R}^{3}.}
\deqn{a(\tau;\theta) = \begin{cases}
\frac{\theta}{\tau} \sin\left( \frac{\tau}{\theta} \right), & \tau \neq = 0 \\
1, & \tau = 0
\end{cases} .
}
}
\examples{
x <- 1
kernel_wave(x, 1)
x <- c(1, 2, 3)
kernel_wave(x, 0.2)
}
\references{
Genton, Marc. (2001). Classes of Kernels for Machine Learning: A Statistics Perspective. Journal of Machine Learning Research. 2. 299-312. 10.1162/15324430260185646.
}
