% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lines.BootEsts}
\alias{lines.BootEsts}
\title{Lines Method for BootEsts Objects}
\usage{
\method{lines}{BootEsts}(x, type = "l", ...)
}
\arguments{
\item{x}{A BootEsts S3 object.}

\item{type}{Defaults to \code{'l'}, see \link[base:plot]{base::plot} for the possible values.}

\item{...}{Additional plotting arguments, refer to \link[graphics:par]{graphics::par}.}
}
\value{
A line of a BootEsts S3 object.
}
\description{
This function plots a BootEsts object as a line onto another plot.
}
\examples{
plot(block_bootstrap(c(1, 2, 3), 2))
lines(block_bootstrap(c(1, 2, 3), 2, pd = FALSE), lwd = 2, lty = 3, col = 'blue')
}
