% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lines.CovEsts}
\alias{lines.CovEsts}
\title{Lines Method for CovEsts Objects}
\usage{
\method{lines}{CovEsts}(x, type = "l", ...)
}
\arguments{
\item{x}{A CovEst S3 object.}

\item{type}{Defaults to \code{'l'}, see \link[base:plot]{base::plot} for the possible values.}

\item{...}{Additional plotting arguments, refer to \link[graphics:par]{graphics::par}.}
}
\value{
A line of a CovEsts S3 object.
}
\description{
This function plots a CovEsts object as a line onto another plot.
}
\examples{
plot(standard_est(c(1, 2, 3)))
lines(standard_est(c(1, 2, 3), pd = FALSE))
}
