% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lines.VarioEsts}
\alias{lines.VarioEsts}
\title{Lines Method for VarioEsts Objects}
\usage{
\method{lines}{VarioEsts}(x, type = "l", ...)
}
\arguments{
\item{x}{A VarioEsts S3 object.}

\item{type}{Defaults to \code{'l'}, see \link[base:plot]{base::plot} for the possible values.}

\item{...}{Additional plotting arguments, refer to \link[graphics:par]{graphics::par}.}
}
\value{
A line of a VarioEsts S3 object.
}
\description{
This function plots a VarioEsts object as a line onto another plot.
}
\examples{
plot(to_vario(standard_est(c(1, 2, 3))))
lines(to_vario(standard_est(c(1, 2, 3), pd = FALSE)))
}
