% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{max_distance}
\alias{max_distance}
\title{The maximum vertical distance between estimated functions.}
\usage{
max_distance(est1, est2, plot = FALSE)
}
\arguments{
\item{est1}{A numeric vector representing the first estimated covariance function.}

\item{est2}{A numeric vector representing the second estimated covariance function.}

\item{plot}{A boolean as to whether a plot should be created. By default, no plot is created.}
}
\value{
A numeric value representing the maximum vertical distance between the two estimated functions.
}
\description{
This function computes the maximum vertical distance between functions.
}
\details{
This function computes the maximum vertical distance between functions. The two functions must be of the same length.
\deqn{D(\hat{C}_{1}(\tau), \hat{C}_{2}(\tau)) = \max_{\tau} \left| \hat{C}_{1}(\tau) - \hat{C}_{2}(\tau) \right| ,
}
where \eqn{\hat{C}_{1}(\cdot)} and \eqn{\hat{C}_{2}(\cdot)} are estimated autocovariance functions.
It assumes that the estimated functions are estimated over the same set of lags.
}
\examples{
x <- seq(0, 5, by=0.1)
estCov1 <- exp(-x^2) + rnorm(length(x), sd=0.1)
estCov2 <- exp(-x^2.1) + rnorm(length(x), sd=0.1)
max_distance(estCov1, estCov2)
max_distance(estCov1, estCov2, plot = TRUE)
}
