% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_est.R
\name{normalise_acf}
\alias{normalise_acf}
\alias{normalise_acf.default}
\alias{normalise_acf.CovEsts}
\title{Normalise a CovEsts Object}
\usage{
normalise_acf(estCov)

\method{normalise_acf}{default}(estCov)

\method{normalise_acf}{CovEsts}(estCov)
}
\arguments{
\item{estCov}{A CovEst S3 object.}
}
\value{
A CovEsts S3 object with a normalised \code{acf} attribute.
}
\description{
This function normalises a CovEsts S3 object, that is, turning an autocovariance function into an autocorrelation function.
This procedure is a one-way transformation.
This function does not support any other type of argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{normalise_acf(default)}: Method for other objects.

\item \code{normalise_acf(CovEsts)}: Method for 'CovEsts' objects.

}}
\examples{
normalise_acf(standard_est(c(1, 2, 3)))
}
