% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.BootEsts}
\alias{plot.BootEsts}
\title{Plot Method for BootEsts Objects}
\usage{
\method{plot}{BootEsts}(x, type = "l", xlab = "Lag (h)", ylab = "ACF", ...)
}
\arguments{
\item{x}{A BootEsts S3 object.}

\item{type}{Defaults to \code{'l'}, see \link[base:plot]{base::plot} for the possible values.}

\item{xlab}{Defaults to \code{'Lag (h)'}, see \link[graphics:title]{graphics::title} for more information.}

\item{ylab}{Defaults to \code{'ACF'}, see \link[graphics:title]{graphics::title} for more information.}

\item{...}{Additional plotting arguments, refer to \link[graphics:par]{graphics::par}.}
}
\value{
A plot of a BootEsts S3 object.
}
\description{
This function plots a BootEsts object. It plots the confidence region, average bootstrap estimate, and the original nonbootstrapped estimate.
}
\examples{
plot(block_bootstrap(c(1, 2, 3), 2))
}
