% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.CovEsts}
\alias{plot.CovEsts}
\title{Plot Method for CovEsts Objects}
\usage{
\method{plot}{CovEsts}(x, type = "l", xlab = "Lag (h)", ylab = "ACF", ...)
}
\arguments{
\item{x}{A CovEst S3 object.}

\item{type}{Defaults to \code{'l'}, see \link[base:plot]{base::plot} for the possible values.}

\item{xlab}{Defaults to \code{'Lag (h)'}, see \link[graphics:title]{graphics::title} for more information.}

\item{ylab}{Defaults to \code{'ACF'}, see \link[graphics:title]{graphics::title} for more information.}

\item{...}{Additional plotting arguments, refer to \link[graphics:par]{graphics::par}.}
}
\value{
A plot of a CovEsts S3 object.
}
\description{
This function plots a CovEsts object, over the lags used to estimate the autocovariance values.
}
\examples{
plot(standard_est(c(1, 2, 3)))
}
