% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.VarioEsts}
\alias{plot.VarioEsts}
\title{Plot Method for VarioEsts Objects}
\usage{
\method{plot}{VarioEsts}(x, type = "l", xlab = "Lag (h)", ylab = expression(gamma(h)), ...)
}
\arguments{
\item{x}{A VarioEsts S3 object.}

\item{type}{Defaults to \code{'l'}, see \link[base:plot]{base::plot} for the possible values.}

\item{xlab}{Defaults to \code{'Lag (h)'}, see \link[graphics:title]{graphics::title} for more information.}

\item{ylab}{Defaults to \code{expression(gamma(h))}, see \link[graphics:title]{graphics::title} for more information.}

\item{...}{Additional plotting arguments, refer to \link[graphics:par]{graphics::par}.}
}
\value{
A plot of a VarioEsts S3 object.
}
\description{
This function plots a VarioEsts object, over the lags used to estimate the variogram values.
}
\examples{
plot(to_vario(standard_est(c(1, 2, 3))))
}
