% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{power_sine_window}
\alias{power_sine_window}
\title{Power Sine Window.}
\usage{
power_sine_window(x, a)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}

\item{a}{A positive number indicating the power.}
}
\value{
A vector of values.
}
\description{
This computes the power sine window.
\deqn{w(x) = \sin^{a}(\pi x / 2), x \in [0, 1], a > 0.}
}
\examples{
X <- c(0.1, 0.5, 0.75)
power_sine_window(X, 3.2)
}
