% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.BootEsts}
\alias{print.BootEsts}
\title{Print Method for BootEsts Objects}
\usage{
\method{print}{BootEsts}(x, n_head = 5, n_tail = 5, digits = NULL, ...)
}
\arguments{
\item{x}{A BootEsts S3 object.}

\item{n_head}{The number of 'head' values to print, defaults to 5.}

\item{n_tail}{The number of 'tail' values to print, defaults to 5.}

\item{digits}{The number of digits to print. Defaults to \code{NULL}. Note, this value is passed into \link[base:format]{base::format}.}

\item{...}{Addition printing parameters, see \link[base:print]{base::print}.}
}
\value{
Prints a BootEsts object.
}
\description{
This function plots a BootEsts object, printing estimtor type, bootstrap type, estimator used, the bootstrap average estimate, the original estimate,
and the pointwise confidence interval.
}
\examples{
print(block_bootstrap(c(1, 2, 3), 2))
}
