% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.CovEsts}
\alias{print.CovEsts}
\title{Print Method for CovEsts Objects}
\usage{
\method{print}{CovEsts}(x, n_head = 5, n_tail = 5, digits = NULL, ...)
}
\arguments{
\item{x}{A CovEst S3 object.}

\item{n_head}{The number of 'head' values to print, defaults to 5.}

\item{n_tail}{The number of 'tail' values to print, defaults to 5.}

\item{digits}{The number of digits to print. Defaults to \code{NULL}. Note, this value is passed into \link[base:format]{base::format}.}

\item{...}{Addition printing parameters, see \link[base:print]{base::print}.}
}
\value{
Prints a CovEsts object.
}
\description{
This function plots a CovEsts object, printing the estimator
type, estimator used, lags and values.
}
\examples{
print(standard_est(c(1, 2, 3)))
}
