% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{rho_T1}
\alias{rho_T1}
\title{Compute \eqn{\rho(T_{1})} for the truncated kernel regression estimator.}
\usage{
rho_T1(
  x,
  meanX,
  T1,
  h,
  xij_mat,
  kernel_name = "gaussian",
  kernel_params = c(),
  custom_kernel = FALSE
)
}
\arguments{
\item{x}{A vector of lags.}

\item{meanX}{The average value of X.}

\item{T1}{The first trunctation point.}

\item{h}{Bandwidth parameter.}

\item{xij_mat}{The matrix of pairwise covariance values.}

\item{kernel_name}{The name of the kernel function to be used. Possible values are:
"gaussian", "wave", "rational_quadratic", and "bessel_j". Alternatively, a custom kernel function can be provided, see \link{compute_corrected_standard_est}'s example.}

\item{kernel_params}{A vector of parameters of the kernel function. See \link{kernel_symm} for parameters.}

\item{custom_kernel}{If a custom kernel is to be used or not.}
}
\value{
The estimated covariance function at \eqn{T_{1}}.
}
\description{
This helper function computes \eqn{\rho(T_{1})} for the truncated kernel regression estimator.
}
\details{
This function computes the following value,
\deqn{
\hat{\rho}(T_{1}) = \left( \sum_{i} \sum_{j} \check{X}_{ij} K((T_{1} - t_{ij}) / h) \right) \left( \sum_{i} \sum_{j} K((T_{1} - t_{ij}) / h) \right)^{-1},
}
where \eqn{\check{X}_{ij} = (X(t_{i}) - \bar{X}) (X(t_{j}) - \bar{X})}, and \eqn{t_{ij} = t_{i} - t_{j}.}
}
\examples{
X <- c(1, 2, 3, 4)
rho_T1(1:4, mean(X), 1, 0.1, Xij_mat(X, mean(X)), "gaussian", c(), FALSE)
}
\references{
Hall, P., & Patil, P. (1994). Properties of Nonparametric Estimators of Autocovariance for Stationary Random Fields. Probability Theory and Related Fields (Vol. 99, Issue 3, pp. 399–424). 10.1007/bf01199899

Hall, P., Fisher, N. I., & Hoffmann, B. (1994). On the nonparametric estimation of covariance functions. The Annals of Statistics (Vol. 22, Issue 4, pp. 2115–2134). 10.1214/aos/1176325774
}
