% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{sine_window}
\alias{sine_window}
\title{Sine Window.}
\usage{
sine_window(x, ...)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}

\item{...}{Extra unusued arguments.}
}
\value{
A vector of values.
}
\description{
This computes the sine window.
\deqn{w(x) = \sin(\pi x / 2), x \in [0, 1].}
}
\examples{
X <- c(0.1, 0.5, 0.75)
sine_window(X)
}
