% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{spectral_norm}
\alias{spectral_norm}
\title{Compute spectral norm between estimated functions.}
\usage{
spectral_norm(est1, est2)
}
\arguments{
\item{est1}{A numeric vector representing the first estimated covariance function.}

\item{est2}{A numeric vector representing the second estimated covariance function.}
}
\value{
The spectral norm of the differences between the two functions.
}
\description{
This function computes the spectral norm between the two estimated autocovariance functions.
}
\details{
This function computes the spectral norm between the two estimated autocovariance functions.
Let \eqn{\hat{D}(\tau) = \hat{C}_{1}(\tau) - \hat{C}_{2}(\tau),}
where \eqn{\hat{C}_{1}(\cdot)} and \eqn{\hat{C}_{2}(\cdot)} are estimated autocovariance functions.

A matrix \eqn{D} is created from \eqn{\hat{D}(\cdot)},
\deqn{\left[ {\begin{array}{ccccc}
D(0)            & D(\tau_{1})     & \cdots & D(\tau_{N - 1}) & D(\tau_{N})     \\
D(\tau_{1})     & D(0)            & \cdots & D(\tau_{N - 2}) & D(\tau_{N - 1}) \\
\vdots          & \vdots          & \ddots & \vdots          & \vdots       \\
D(\tau_{N - 1}) & D(\tau_{N - 2}) & \cdots & D(0)            & D(\tau_{1})     \\
D(\tau_{N})     & D(\tau_{N - 1}) & \cdots & D(\tau_{1})     & D(0)         \\
\end{array}} \right] ,
}
over a set of lags \eqn{\{0, \tau_{1}, \dots , \tau_{N} \}.}
This matrix is created in \link{create_cyclic_matrix}.

The spectral norm is simply the largest singular value of \eqn{D.}
}
\examples{
x <- seq(0, 5, by=0.1)
estCov1 <- exp(-x^2) + rnorm(length(x), sd=0.1)
estCov2 <- exp(-x^2.1) + rnorm(length(x), sd=0.1)
spectral_norm(estCov1, estCov2)
}
