% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_est.R
\name{standard_est}
\alias{standard_est}
\title{Computes the autocovariance function using the standard estimator.}
\usage{
standard_est(X, upperTau, N = length(X), meanX = mean(X), pd = TRUE)
}
\arguments{
\item{X}{A vector of observations representing a time series.}

\item{upperTau}{The maximum upper lag to compute the autocovariance function. upperTau < N-1.}

\item{N}{The length of the vector X.}

\item{meanX}{The average value of the observations.}

\item{pd}{Whether a positive definite estimate should be used.}
}
\value{
The autocovariance up to lag upperTau for the time series X.
}
\description{
This computes the following two variants of the autocovariance function depending on
the parameter \code{pd}.
}
\details{
For \code{pd = TRUE}:
\deqn{
\widehat{C}(\tau) = \frac{1}{N} \sum_{i=1}^{N-\tau} ( X(t_{i}) - \bar{X} ) ( X(t_{i} + \tau) - \bar{X} ) .
}

For \code{pd = FALSE}:
\deqn{
\widehat{C}(\tau) = \frac{1}{N - \tau} \sum_{i=1}^{N-\tau} ( X(t_{i}) - \bar{X} ) ( X(t_{i} + \tau) - \bar{X} ) .
}

This function will generate autocovariance values for lags \eqn{\{0, \dots, upperTau\}.}
}
\examples{
X <- c(1, 2, 3)
standard_est(X, 2, length(X), mean(X), pd=FALSE)
}
