% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{triangular_window}
\alias{triangular_window}
\title{Triangular Window.}
\usage{
triangular_window(x)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}
}
\value{
A vector of values.
}
\description{
This computes the triangular window.
\deqn{w(x) = x, x \in [0, 1].}
}
\examples{
X <- c(0.1, 0.5, 0.75)
triangular_window(X)
}
