% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{tukey_window}
\alias{tukey_window}
\title{Tukey Window.}
\usage{
tukey_window(x, ...)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}

\item{...}{Extra unusued arguments.}
}
\value{
A vector of values.
}
\description{
This function computes the Tukey window.
\deqn{w(x) = \frac{1}{2} \cos(\pi x) , x \in [0, 1].}
}
\examples{
X <- c(0.1, 0.5, 0.75)
tukey_window(X)
}
