% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{welch_window}
\alias{welch_window}
\title{Welch Window.}
\usage{
welch_window(x, ...)
}
\arguments{
\item{x}{A vector of values at which the window is computed at.}

\item{...}{Extra unusued arguments.}
}
\value{
A vector of values.
}
\description{
This computes the Welch  window.
\deqn{w(x) = 1 - (x - 1)^2 , x \in [0, 1] . }
}
\examples{
X <- c(0.1, 0.5, 0.75)
welch_window(X)
}
