% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{window}
\alias{window}
\title{1D window functions.}
\usage{
window(x, name, params = c(1))
}
\arguments{
\item{x}{A vector or matrix of values of at least length 1.}

\item{name}{The name of the window. Options are: "tukey", "triangular", "sine", "power_sine", "blackman", "hann_poisson", "welch".}

\item{params}{A vector of parameters for the windows. See the documentation below for the position of the parameters.}
}
\value{
A vector of values.
}
\description{
A window function in this context is a continuous nondecreasing function such that at 0 it is 0, and at 1, it is 1.
This computes one of the window functions listed below.
}
\details{
\strong{Tukey Window}.
\deqn{w(x) = \frac{1}{2} - \frac{1}{2} \cos(\pi x) , x \in [0, 1].}
The \code{params} argument is empty, see the example.

\strong{Triangular Window}.
\deqn{w(x) = x, x \in [0, 1].}
The \code{params} argument is empty, see the example.

\strong{Sine Window}.
\deqn{w(x) = \sin\left(\pi x / 2 \right), x \in [0, 1].}
The \code{params} argument is empty, see the example.

\strong{Power Sine Window}.
\deqn{w(x; a) = \sin^{a}(\pi x / 2), x \in [0, 1], a > 0.}
The \code{params} argument is of the form \code{c(}\eqn{a}\code{)}

\strong{Blackman Window}.
\deqn{w(x; \alpha) = ( (1 - \alpha) / 2) - \frac{1}{2} \cos(\pi x) + \frac{\alpha}{2} \cos(2 \pi x), x \in [0, 1], \alpha \in {R} .}
The \code{params} argument is of the form \code{c(}\eqn{\alpha}\code{)}
It is recommended that \eqn{\alpha = 0.16} to ensure that \eqn{w(0) = 0, w(1) = 1.}

\strong{Hann-Poisson Window}.
\deqn{w(x; \alpha) = \frac{1}{2} (1 - \cos(\pi x)) \exp( - (\alpha \left|1 - x \right|) ) , x \in [0, 1], \alpha \in {R} .}
The \code{params} argument is of the form \code{c(}\eqn{\alpha}\code{)}

\strong{Welch Window}.
\deqn{w(x) = 1 - (x - 1)^2 , x \in [0, 1] . }
The \code{params} argument is empty, see the example.
}
\examples{
x <- c(0.2, 0.4, 0.6)
window(x, "tukey")
window(x, "triangular")
window(x, "sine")
window(x, "power_sine", c(0.7))
window(x, "blackman", c(0.16))
window(x, "hann_poisson", c(0.7))
window(x, "welch")
}
