% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{block_bootstrap}
\alias{block_bootstrap}
\title{Title}
\usage{
block_bootstrap(
  X,
  maxLag,
  n_bootstrap = 100,
  l = ceiling(length(X)^(1/3)),
  estimator = standard_est,
  p_quantile = 0.95,
  plot = FALSE,
  ...
)
}
\arguments{
\item{X}{A vector representing observed values of the time series.}

\item{maxLag}{The maximum lag to compute the estimated autocovariance function at.}

\item{n_bootstrap}{The number of times to run moving block bootstrap. Defaults to 100.}

\item{l}{The block length considered for bootstrap. Defaults to \eqn{\ceiling{N}^{1/3}}, where \eqn{N} is the length of the observation window.}

\item{estimator}{The function name of the estimator to use. Defaults to \code{standard\_est}.}

\item{p_quantile}{The quantile used to compute the \eqn{(1 - \alpha)\%} confidence interval. Defaults to \eqn{0.95.}}

\item{plot}{A boolean determining whether a plot should be created. By default, no plot is created.}

\item{...}{Optional arguments to the chosen estimator. See the examples.}
}
\value{
A list consisting of three items. The f irst is the average estimated autocovariance/autocorrelation function for the bootstrap samples, the second is a matrix of the estimated autocovariance/autocorrelation functions from the bootstrapped samples, and the third is a matrix of confidence intervals for each lag.
}
\description{
Title
}
\examples{
X <- c(1, 2, 3, 3, 2, 1)
block_bootstrap(X, 4, n_bootstrap = 100, l = 2, type = 'autocorrelation')
block_bootstrap(X, 4, n_bootstrap = 100, l = 2, plot =TRUE, type = 'autocovariance')
block_bootstrap(X, 4, n_bootstrap = 100, l = 2, estimator=tapered_est, rho = 0.5, window_name = 'blackman', window_params = c(0.16), type='autocorrelation')
}
