% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Moving Block Bootstrap Samples}
\usage{
bootstrap_samples(X, l, k, boot_type = "moving")
}
\arguments{
\item{X}{A vector representing observed values of the time series.}

\item{l}{The block length considered for bootstrap.}

\item{k}{The number of blocks considered for bootstrap.}

\item{boot_type}{What type of block bootstrap should be used, either 'moving' for moving block bootstrap or 'circular' for circular block bootstrap.}
}
\value{
A vector of length \code{length(X)} whose values are a bootstrapped time series.
}
\description{
This function generates moving block bootstrap samples.
}
\details{
This function generates a moving block bootstrap sample for a time series \eqn{X.}
For some times series \eqn{X_{1}, X_{2}, \dots, X_{n},} construct \eqn{k \in Z} overlapping blocks of the form
\eqn{B_{i} = (X_{i}, \dots, X_{i + \ell - 1}),} where \eqn{\ell \in [1, n]} is the block length.
Randomly sample, with replacement, from the discrete uniform distribution with on \eqn{\{1, \dots, n - \ell + 1\}} to obtain a set of random block locations \eqn{I_{1}, \dots, I_{k}.}
Construct a bootstrapped time series \eqn{X_{1}^{\ast}, X_{2}^{\ast}, \dots, X_{n}^{\ast}} where \eqn{X_{i}^{\ast} = B_{I_{i}}.}
The bootstrapped time series is truncated to have an observation window of size \eqn{n.}
}
\examples{
X <- c(1, 2, 3, 3, 2, 1)
bootstrap_samples(X, 2, 3)
}
\references{
Chapters 2.5 and 2.7 in Lahiri, S. N. (2003). Resampling Methods for Dependent Data. Springer. https://doi.org/10.1007/978-1-4757-3803-2

Künsch, H. R. (1989). The Jackknife and the Bootstrap for General Stationary Observations. The Annals of Statistics 17(3), 1217-1241. https://doi.org/10.1214/aos/1176347265

Politis, D. N. & Romano, J. P. (1991). A Circular Block-Resampling Procedure for Stationary Data. In R. LePage & L. Billard, eds, Exploring the Limits of Bootstrap, Wiley, 263-270.
}
