% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{compute_adjusted_est}
\alias{compute_adjusted_est}
\title{Compute the Kernel Regression Estimator.}
\usage{
compute_adjusted_est(
  X,
  x,
  t,
  h,
  kernel_name = "gaussian",
  kernel_params = c(),
  custom_kernel = FALSE,
  pd = TRUE,
  type = "covariance",
  meanX = mean(X)
)
}
\arguments{
\item{X}{A vector representing the values of the process.}

\item{x}{A vector of lags.}

\item{t}{The values at which the covariance function is calculated at.}

\item{h}{Bandwidth parameter.}

\item{kernel_name}{The name of the kernel function to be used. Possible values are:
"gaussian", "wave", "rational_quadratic", and "bessel_j". Alternatively, a custom kernel function can be provided, see \link{compute_corrected_standard_est}'s example.}

\item{kernel_params}{A vector of parameters of the kernel function. See \link{kernel_symm} for parameters.}

\item{custom_kernel}{If a custom kernel is to be used or not.}

\item{pd}{Whether a positive-definite estimate should be used. Defaults to \code{TRUE}.}

\item{type}{Compute either the 'covariance' or 'correlation'. Defaults to 'covariance'.}

\item{meanX}{The average value of X. Defaults to \code{mean(X)}.}
}
\value{
A vector whose values are the kernel regression estimator.
}
\description{
The kernel regression estimator is defined as
\deqn{
\hat{\rho}(t) = \left( \sum_{i} \sum_{j} \check{X}_{ij} K((t - t_{ij}) / h) \right) \left( \sum_{i} \sum_{j} K((t - t_{ij}) / h) \right)^{-1},
}
where \eqn{\check{X}_{ij} = (X(t_{i}) - \bar{X}) (X(t_{j}) - \bar{X})}, and \eqn{t_{ij} = t_{i} - t_{j}.}
}
\details{
The kernel regression estimator of an autocovariance function is defined as
\deqn{
\hat{\rho}(t) = \left( \sum_{i} \sum_{j} \check{X}_{ij} K((t - t_{ij}) / h) \right) \left( \sum_{i} \sum_{j} K((t - t_{ij}) / h) \right)^{-1},
}
where \eqn{\check{X}_{ij} = (X(t_{i}) - \bar{X}) (X(t_{j}) - \bar{X})}, and \eqn{t_{ij} = t_{i} - t_{j}.}

If \code{pd} is \code{TRUE} estimator will be made positive-definite through the following procedure
\enumerate{
\item Take the discrete cosine transform,
\eqn{\mathcal{F}^{c}(\hat{\rho}(t))}.
\item Set all negative values to zero,
\eqn{\hat{\mathcal{F}}^{c}(\hat{\rho}(t)) = \mathcal{F}(\hat{\rho}(t))} if \eqn{\mathcal{F}(\hat{\rho}(t)) > 0} and \eqn{\hat{\mathcal{F}}^{c}(\hat{\rho}(t)) = 0} otherwise,
for all sample frequencies.
\item Perform the inversion to obtain a new estimator.
}

This ensures the covariance estimate is positive-definite.
}
\examples{
X <- c(1, 2, 3, 4)
compute_adjusted_est(X, 1:4, 1:3, 0.1, "gaussian", c(), FALSE, TRUE)
}
\references{
Hall, P., & Patil, P. (1994). Properties of nonparametric estimators of autocovariance for stationary random fields. Probability Theory and Related Fields (Vol. 99, Issue 3, pp. 399–424). 10.1007/bf01199899

Hall, P., Fisher, N. I., & Hoffmann, B. (1994). On the nonparametric estimation of covariance functions. The Annals of Statistics (Vol. 22, Issue 4, pp. 2115–2134). 10.1214/aos/1176325774
}
