% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapered_estimator.R
\name{compute_tapered_cov}
\alias{compute_tapered_cov}
\title{Compute the tapered covariance function over a set of lags.}
\usage{
compute_tapered_cov(
  X,
  maxLag,
  rho,
  window_name,
  window_params = c(1),
  custom_window = FALSE
)
}
\arguments{
\item{X}{A numeric vector representing the process.}

\item{maxLag}{The maximum lag at which the covariance function is to be computed. Must be less than the length of the process.}

\item{rho}{A scale parameter greater than zero.}

\item{window_name}{The window function to use, one of
"tukey", "triangular", "power_sine", "blackman_window", "hann_poisson", "welch".}

\item{window_params}{Any parameters of the window function.}

\item{custom_window}{Whether or not a custom window is used (boolean). Put its parameters in window_param}
}
\value{
A numeric vector representing values of the tapered covariance function.
}
\description{
This computes the tapered covariance over a set of lags,
\deqn{\widehat{C}_{N}^{a} (h) = (H_{2, n})^{-1} \sum_{t=1}^{N-h}  (X(t_{i}) - \bar{X} ) ( X(t_{i} + h) - \bar{X} ) [ a((t_{i} - 1/2) / n; \rho) a((t_{i} + h - 1/2) / n; \rho)  ] ,}
where \eqn{a} is a window function, \eqn{\rho} is a scale parameter.
For each lag, the tapered covariance is computed in \link{compute_tapered_cov_single}.
}
\examples{
X <- c(1, 2, 3)
compute_tapered_cov(X, 2, 0.5, "tukey")
}
