% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_regression_estimator.R
\name{dct_1d}
\alias{dct_1d}
\title{Compute 1D Discrete Cosine Transform}
\usage{
dct_1d(X)
}
\arguments{
\item{X}{A vector of values for which the discrete cosine transform is being computed.}
}
\value{
A vector of discrete cosine transform values.
}
\description{
This function computes the Type-II discrete cosine transform.
}
\details{
The Type-II discrete cosine transform is obtained using \link[stats:fft]{stats::fft}. Note this method does not utilise Makhoul's method for computing the DCT.
A new input \eqn{y} is created from \eqn{X} where \eqn{y_{2n} = 0}, \eqn{y_{2n+1} = X_{n}} for \eqn{0 \le n < N}, \eqn{y_{2N} = 0} and \eqn{y_{4N - n} = y_{n}}
for \eqn{0 < n < 2N}, where N is the length of \eqn{X}.
After this, the is done (in code): \code{dct <- 0.5 * Re(stats::fft(Y))[1:(length(Y) / 4)]}, which gives the Type-II discrete cosine transform.
}
\examples{
X <- c(1, 2, 3)
dct_1d(X)
}
\references{
Ochoa-Dominguez, H., & Rao, K.R. (2019). Discrete Cosine Transform, Second Edition. CRC Press. 10.1201/9780203729854
}
