% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_splines_est.R
\name{get_all_tau}
\alias{get_all_tau}
\title{Get all \eqn{\tau}.}
\usage{
get_all_tau(p, m)
}
\arguments{
\item{p}{The order of the splines.}

\item{m}{The number of nonboundary knots.}
}
\value{
A numerical vector representing all \eqn{\tau_{i}}s.
}
\description{
Repeatedly call \link{get_tau} for all \eqn{\tau_{i}.}
See Choi, Li & Wang (2013, p. 615) for details.
}
\examples{
get_all_tau(3, 2)
}
\references{
Choi, I., Li, B. & Wang, X. Nonparametric Estimation of Spatial and Space-Time Covariance Function. JABES 18, 611–630 (2013). https://doi.org/10.1007/s13253-013-0152-z
}
