% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_splines_est.R
\name{get_splines_df}
\alias{get_splines_df}
\title{Construct dataframe of basis functions.}
\usage{
get_splines_df(x, p, m, taus)
}
\arguments{
\item{x}{A vector of x values - the lags.}

\item{p}{The order of splines.}

\item{m}{The number of nonboundary knots.}

\item{taus}{Vector of \eqn{\tau}s, see \link{get_tau}.}
}
\value{
A dataframe of the structure found above.
}
\description{
This function constructs a dataframe with the following structure:
\itemize{
\item One column for the x-values
\item m + p columns of squared basis functions evaluated at the correspond x.
}
}
\examples{
taus <- get_all_tau(3, 2)
get_splines_df(seq(0, 2, by=0.25), 3, 2, taus)
}
