% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_splines_est.R
\name{get_tau}
\alias{get_tau}
\title{Get a specific \eqn{\tau_{i}}.}
\usage{
get_tau(i, p, m, kVec)
}
\arguments{
\item{i}{The knot number (\eqn{-p} through \eqn{m + p + 1}).}

\item{p}{The order of splines.}

\item{m}{The number of nonboundary knots.}

\item{kVec}{Knot vector - see \link{generate_knots}.}
}
\value{
A numerical value representing \eqn{\tau_{i}.}
}
\description{
Transforms the knots from \link{generate_knots} into the following form:
For \eqn{i = -p , \dots , -1 , m + 2, \dots , m + p + 1, \tau_{i} = i / (m + 1)}, and for \eqn{i = 0, \dots , m + 1, \tau_{i} = \kappa_{i}.}
See Choi, Li & Wang (2013) page 615 for details.
This is a helper function of \link{get_all_tau}.
}
\examples{
kVec <- generate_knots(2)
get_tau(1, 3, 2, kVec)
}
\references{
Choi, I., Li, B. & Wang, X. Nonparametric Estimation of Spatial and Space-Time Covariance Function. JABES 18, 611–630 (2013). https://doi.org/10.1007/s13253-013-0152-z
}
