% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nearest_pd}
\alias{nearest_pd}
\title{Compute the Nearest Positive-Definite Matrix.}
\usage{
nearest_pd(X, return_matrix = FALSE)
}
\arguments{
\item{X}{Either a numeric vector or a square matrix. If a vector is provided, a matrix will be created of the form found in \link{create_cyclic_matrix}.}

\item{return_matrix}{A boolean determining whether or not the closest matrix is given or not. If \code{FALSE}, it returns the closest autocorrelation function. Defaults to \code{FALSE}.}
}
\value{
The closest positive-definite autocorrelation function or matrix (depending on \code{return_matrix}).
}
\description{
This function computes the nearest positive-definite matrix to some matrix \eqn{X}.
}
\examples{
X <- c(1, 0, -1.1)
nearest_pd(X)
nearest_pd(X, TRUE)
}
\references{
Higham, N. J. (1988). Computing a nearest symmetric positive semidefinite matrix. Linear Algebra and its Applications, 103, 103–118. 10.1016/0024-3795(88)90223-6
}
