% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_splines_est.R
\name{solve_spline}
\alias{solve_spline}
\title{Objective Function for WLS.}
\usage{
solve_spline(par, splines_df, weights)
}
\arguments{
\item{par}{A vector of parameters to minimise.}

\item{splines_df}{A dataframe whose structure is found in \link{get_splines_df}.}

\item{weights}{A vector of weights, see the description.}
}
\value{
The value of the objective function at those parameters.
}
\description{
This is the objective function to find the weights for each basis function, see Choi, Li & Wang (2013, p. 617).
The parameters must be nonnegative, so a penalty of \eqn{10^{12}} is given if any parameters are negative.
The weights are chosen as per Choi, Li & Wang (2013, p. 617).
}
\details{
Let \eqn{\mathbf{\beta} = (\beta_{0}, \dots, \beta_{m + p})^{\prime}} be a vector of model coefficients, \eqn{\{f_{1}^{(p - 1)} , \dots , f_{m + p}^{(p - 1)} \}}
are a set of completely monotone basis functions, and \eqn{\widehat{C}} is an estimated covariance function.
As per Choi, Li & Wang (2013, p. 617), \eqn{\mathbf{\beta}} can be estimated via weighted-least squares,
\deqn{
\hat{\mathbf{\beta}}_{WLS} = {\arg\min}_{\beta_{j} \ge 0} \sum_{i=1}^{L} w_{i} \left(\widehat{C}(\tau_{i}) - \sum_{j = 1}^{m + p} \beta_{j} f_{j}^{(p - 1)}(\tau_{i}^{2})  \right)^{2}
},
where there is a set of lags \eqn{\{\tau_{1} , \dots , \tau_{L} \}} and a set of weights \eqn{\{w_{1}, \dots , w_{L} \}.}
The set of weights are calculated in \link{splines_est}, and are of form
\eqn{w_{i} = (N - \tau_{i}) / ((1 - \widehat{C}(\tau_{i}))^{2}).}
}
\examples{
taus <- get_all_tau(3, 2)
x <- seq(0, 2, by=0.25)
maxLag <- 4
splines_df <- get_splines_df(x[1:maxLag], 3, 2, taus)
splines_df$estCov <- exp(-splines_df$lags^2)
pars <- c(0.5, 0.5, 0.5, 0.5, 0.5)
weights <- c()
X <- rnorm(50)
for(i in 0:(maxLag - 1)) {
weights <- c(weights, (length(X) - i) / ( (1 - splines_df$estCov[i + 1])^2 ))
}
solve_spline(pars, splines_df, weights)
}
\references{
Choi, I., Li, B. & Wang, X. Nonparametric Estimation of Spatial and Space-Time Covariance Function. JABES 18, 611–630 (2013). https://doi.org/10.1007/s13253-013-0152-z
}
