% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{starting_locs}
\alias{starting_locs}
\title{Random Block Locations}
\usage{
starting_locs(N, l, k)
}
\arguments{
\item{N}{The length of the observation window.}

\item{l}{The block length considered for bootstrap.}

\item{k}{The number of blocks considered for bootstrap.}
}
\value{
A vector of length \code{k} whose values are random
}
\description{

}
\examples{
starting_locs(4, 2, 2)
}
