% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_est.R
\name{to_vario}
\alias{to_vario}
\title{Autocovariance to Semivariogram}
\usage{
to_vario(cov)
}
\arguments{
\item{cov}{A vector whose values are an estimate autocovariance function.}
}
\value{
A vector whose values are an approximate estimate to the semivariogram.
}
\description{
This function computes an approximate estimated semivariogram from an an estimated autocovariance function.
}
\details{
The semivariogram, \eqn{\gamma(h)} and autocovariance function, \eqn{C(h)}, under the assumption of weak stationarity are related as follows:
\deqn{
\gamma(h) = C(0) - C(h) .
}

When an empirical autocovariance function is considered instead, this relation does not necessarily hold, however,
it can be used to obtain a function that is close to a semivariogram.
}
\examples{
X <- c(1, 2, 3)
estCov <- compute_standard_est(X, 2, length(X), mean(X), pd=FALSE)
to_vario(estCov)
}
